# support set --env-file

+++ server/init/main.py
@@ -2,6 +2,7 @@
 import argparse
 import logging
 import pathlib
+import sys
 
 import share
 import yaml
@@ -17,20 +18,26 @@ logger = logging.getLogger()
 
 if __name__ == '__main__':
     log_util.init_logger()
-    env_dict = yaml_util.load(file_util.read_text(pathlib.Path(__file__).parent.joinpath("env.yaml")))
-    file_util.write_text(pathlib.Path(__file__).parent.joinpath("vars/env.yml"), yaml.dump(env_dict))
-
     private_key = pathlib.Path(__file__).parent.joinpath("___temp/private-key").as_posix()
     ansible_hosts = pathlib.Path(__file__).parent.joinpath("ansible-hosts").as_posix()
     parser = argparse.ArgumentParser(formatter_class=share.CustomHelpFormatter, conflict_handler="resolve")
     share.Installer.set_common_argument(parser)
+    parser.add_argument('--env-file', required=False, default="env.yaml", type=str, help="env file xxx.yaml")
     parser.add_argument('-f', '--file', required=True, type=str, help="inventory file")
     parser.add_argument('-t', '--tag', required=True, type=str, help="t1,t2")
     parser.add_argument('-k', '--ask-pass', action="store_true", help="ask for connection password")
-    parser.add_argument('-u', '--user', default=env_dict["param_user_ops"], type=str, help="connect as this user (default=[param_user_ops])")
+    parser.add_argument('-u', '--user', required=False, type=str, help="connect as this user (default=[param_user_ops])")
     parser.add_argument('--check', action="store_true", help="don't make any changes")
     parser.add_argument('--no-step', action="store_true", help="disable one-step-at-a-time")
     args = parser.parse_args()
+    env_file: pathlib.Path = pathlib.Path(__file__).parent.joinpath(args.env_file)
+    if not env_file.exists():
+        logger.error("env file not exists")
+        sys.exit(0)
+    env_dict = yaml_util.load(file_util.read_text(env_file))
+    file_util.write_text(pathlib.Path(__file__).parent.joinpath("vars/env.yml"), yaml.dump(env_dict))
+    if not args.user:
+        args.user = env_dict["param_user_ops"]
     if args.debug:
         logger.setLevel(logging.DEBUG)
